separate explore area
